Rename firabaseActions => workActions.
