+++ src/states/work.tsx
@@ -1,6 +1,6 @@
 import { Success } from 'typescript-fsa';
 import { reducerWithInitialState } from 'typescript-fsa-reducers';
-import { firebaseActions, IWork } from '../actions/actions';
+import { workActions, IWork } from '../actions/actions';
 
 export interface IWorkState extends IWork {
     thumbnailURL: string
@@ -27,13 +27,13 @@ const initialFirebaseState: IFirebaseState = {
 };
 
 export const firebaseReducer = reducerWithInitialState(initialFirebaseState)
-    .case(firebaseActions.getWorks.started, (state: IFirebaseState) => {
+    .case(workActions.getWorks.started, (state: IFirebaseState) => {
         return Object.assign({}, state, { isLoading: true });
     })
-    .case(firebaseActions.getWorks.done, (state: IFirebaseState, success: Success<void, IWork[]>) => {
+    .case(workActions.getWorks.done, (state: IFirebaseState, success: Success<void, IWork[]>) => {
         return Object.assign({}, state, { works: success.result });
     })
-    .case(firebaseActions.getThumbnail.done, (state: IFirebaseState, success: Success<string, string>) => {
+    .case(workActions.getThumbnail.done, (state: IFirebaseState, success: Success<string, string>) => {
         let stateTmp = Object.assign(state);
         stateTmp.isLoading = false;
         stateTmp.works
@@ -42,7 +42,7 @@ export const firebaseReducer = reducerWithInitialState(initialFirebaseState)
 
         return Object.assign({}, state, { stateTmp });
     })
-    .case(firebaseActions.getDetailImages.done, (state: IFirebaseState, success: Success<IWorkState | null, string[]>) => {
+    .case(workActions.getDetailImages.done, (state: IFirebaseState, success: Success<IWorkState | null, string[]>) => {
         if (success.params !== null) {
             const title = success.params.title
             let stateTmp = Object.assign(state);
